requirements.txt
grpcio==1.73.1
grpcio-tools==1.73.1
kafka-python==2.2.15
protobuf==6.31.1
python-dotenv==1.1.1
